+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.3.0
+* Rewrite of monitor_device function
+* Fix leak of File in log_message
+* Improved commenting
+* Implemented shared memory requests without using signals 
+
 0.2.2
 * PATCH Shared memory update after cooldown
 * Fix a typos in log output and function name
@@ -12,10 +18,11 @@
 0.2.0
 * Added shared memory interface
 * Changed Logging system
-* Added CLI program to interface with daemon
+* Added CLI program to interface with daemon  
  
-
 0.1.6  
 * fix signed unsigned conversion  
 * fix unexpected exit if pulse width out of bounds  
-0.1.5 Initial Release
+
+0.1.5
+* Initial Release